Optimize squashfs image for live media
Removing the intermediate rootfs.img let us save space (at the time of
this writing it's 200 MB with a Liri OS image) and simplify not only the
build process but also the boot.

This has proven to also fix boot issue reported in #3.

We also pass different arguments because we want to replace the
dmsquash-live dracut module with our own implementation, that doesn't
deal with the persistent storage.

However the squashfs image under LiveOS/ could probably be booted by
dmsquash-live, so in the future we might want to make this configurable.

The ostree-live dracut module that implements our boot process is
currently in the `ostree-config` repository under `lirios`.

Closes: #3
Closes: #4