Rewrite with Go and overhaul the architecture
Go is easier to write and maintain, builds faster and has finally
improvements on how modules are managed.

Instead of rewriting the code to do exactly what it previously did, the
architecture gets an overhaul.

Recipes are now declarative: the actions are going to be executed one
after another.  You can also use variables that are expanded with Go
template library.

Users are also free to provide their bootloader configuration templates.